+++ test/topology/test_topology_failure_recovery.py
@@ -74,28 +74,3 @@ async def test_topology_streaming_failure(request, manager: ManagerClient):
     assert s not in servers
     matches = [await log.grep("storage_service - rollback.*after replacing failure to state left_token_ring", from_mark=mark) for log, mark in zip(logs, marks)]
     assert sum(len(x) for x in matches) == 1
-
-@pytest.mark.asyncio
-async def test_tablet_drain_failure_during_decommission(manager: ManagerClient):
-    servers = await manager.running_servers()
-
-    logs = [await manager.server_open_log(srv.server_id) for srv in servers]
-    marks = [await log.mark() for log in logs]
-
-    cql = manager.get_cql()
-    await cql.run_async("CREATE KEYSPACE test WITH replication = {'class': 'NetworkTopologyStrategy', 'replication_factor': 1} AND tablets = {'initial': 32};")
-    await cql.run_async("CREATE TABLE test.test (pk int PRIMARY KEY, c int);")
-
-    logger.info("Populating table")
-
-    keys = range(256)
-    await asyncio.gather(*[cql.run_async(f"INSERT INTO test.test (pk, c) VALUES ({k}, {k});") for k in keys])
-
-    await inject_error_on(manager, "stream_tablet_fail_on_drain", servers)
-
-    await manager.decommission_node(servers[2].server_id, expected_error="Decommission failed. See earlier errors")
-
-    matches = [await log.grep("storage_service - rollback.*after decommissioning failure to state rollback_to_normal", from_mark=mark) for log, mark in zip(logs, marks)]
-    assert sum(len(x) for x in matches) == 1
-
-    await cql.run_async("DROP KEYSPACE test;")
