+++ test/topology_custom/suite.yaml
@@ -17,6 +17,7 @@ skip_in_release:
   - test_old_ip_notification_repro
   - test_different_group0_ids
   - test_group0_schema_versioning
+  - test_topology_failure_recovery
 skip_in_debug:
   - test_shutdown_hang
   - test_replace
